feat: move BusFile to storagebus/storage package
