--- packages/file/CHANGELOG.md
@@ -1,1 +1,0 @@
-# @storagebus/file
