--- packages/file/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@storagebus/file",
-  "version": "0.1.0",
-  "description": "",
-  "license": "MIT",
-  "author": "Ducktors <your@ducktors.dev> (https://ducktors.dev)",
-  "private": false,
-  "files": ["dist"],
-  "type": "commonjs",
-  "exports": {
-    ".": {
-      "typescript": "./src/file.ts",
-      "import": "./dist/file.js",
-      "require": "./dist/file.js"
-    },
-    "./errors": {
-      "typescript": "./src/errors.ts",
-      "import": "./dist/errors.js",
-      "require": "./dist/errors.js"
-    }
-  },
-  "main": "dist/file.js",
-  "scripts": {
-    "build": "rimraf dist && tsc",
-    "format": "biome format --write .",
-    "lint": "biome check .",
-    "preinstall": "npx only-allow pnpm",
-    "prepublish": "pnpm build",
-    "test": "tsx --conditions=typescript --test ./test/*.ts",
-    "test:ci": "c8 --all --src dist --reporter lcov --reporter text tsx --test ./test/*.ts",
-    "test:watch": "tsx --conditions=typescript --watch --test ./test/*.ts"
-  },
-  "keywords": [],
-  "dependencies": {
-    "tslib": "^2.6.2"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "^1.2.2",
-    "@ducktors/tsconfig": "^1.0.0",
-    "@types/node": "^20.6.5",
-    "c8": "^8.0.1",
-    "rimraf": "^5.0.1",
-    "tsx": "^3.13.0",
-    "typescript": "^5.2.2"
-  }
-}
