--- packages/file/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@ducktors/tsconfig",
-  "compilerOptions": {
-    "outDir": "dist",
-    "sourceMap": true,
-    "declarationMap": true,
-    "composite": true,
-    "rootDir": "src",
-    "tsBuildInfoFile": "dist/tsconfig.build.tsbuildinfo"
-  },
-  "include": ["src", "src/**/*.json"]
-}
