+++ packages/local/src/local.ts
@@ -1,15 +1,13 @@
-import { createReadStream, createWriteStream } from 'node:fs'
-
-import { mkdir, unlink, stat } from 'node:fs/promises'
-import { dirname, join } from 'node:path'
-import { pipeline } from 'node:stream/promises'
-
 import {
   StorageOptions as StorageBusOptions,
   Storage as StorageBus,
   Driver,
 } from '@storagebus/storage'
-import { ENOENT } from '@storagebus/storage/file'
+import { ENOENT } from '@storagebus/storage/errors'
+import { createReadStream, createWriteStream } from 'node:fs'
+import { mkdir, unlink, stat } from 'node:fs/promises'
+import { dirname, join } from 'node:path'
+import { pipeline } from 'node:stream/promises'
 import { once } from 'node:stream'
 
 export type StorageOptions = {
