+++ packages/storage/package.json
@@ -18,6 +18,11 @@
       "import": "./dist/file.js",
       "require": "./dist/file.js"
     },
+    "./errors": {
+      "typescript": "./src/errors.ts",
+      "import": "./dist/errors.js",
+      "require": "./dist/errors.js"
+    },
     "./memory": {
       "typescript": "./src/memory.ts",
       "import": "./dist/memory.js",
@@ -47,7 +52,6 @@
   },
   "keywords": [],
   "dependencies": {
-    "@storagebus/file": "workspace:*",
     "mime-types": "^2.1.35",
     "tslib": "^2.6.2"
   },
